requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.8.0
fastapi==0.110.2
filelock==3.13.4
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.23.2
packaging==24.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
sympy==1.12
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.11.0
uvicorn==0.29.0